fix: move PaginationData class from Deprecated namespace
