+++ NjordinSight.Web/Controllers/ICollectionController.cs
@@ -1,5 +1,5 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using NjordinSight.DataTransfer.Deprecated;
+using NjordinSight.DataTransfer;
 using NjordinSight.EntityModelService.Abstractions;
 using System;
 using System.Collections.Generic;
